Revert "279 move PropertyNameAttribute to separate namespace"
This reverts commit 02809972d04641e6678f305d186e27c05047a7e8.
