+++ Kentico.Kontent.Delivery/ContentItems/PropertyNameAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Kentico.Kontent.Delivery.ContentItems.Attributes
+namespace Kentico.Kontent.Delivery.ContentItems
 {
     /// <summary>
     /// Instructs the <see cref="PropertyMapper"/> to always map to the property with the specified name.
